Overhauled DTOs (part 1)
Now, right before 4.0, is the perfect time to clean up our DTOs, since
this will change some return types or even class names.

- Checked for each endpoint, which variables are documented in the Riot
Api Doc
- Changed some return types even if the Riot Api Doc states differently.
For example, championIds are now always int, instead of sometimes being
considered long
- Deprecated all non-listed variables and their get methods
- Removed some existing setMethods for DTOs
- Renamed some DTOs according to Riot Api Doc (e.g. Participant ->
CurrentGameParticipant; this should also help for clarity where both,
Match.Participant and CurrentGame.Participant are used)
- Reordered DTOs:
- - Variables are ordered by their appearance in the Riot Api Doc. This
helps to check for updated variables at a later point
- - Methods are ordered alphabetically

(for now the endpoints "champion", "championmastery" and "current_game"
are done. will do the rest tomorrow, i'm on the run right now)

+++ src/main/java/net/rithms/riot/api/endpoints/current_game/dto/Rune.java
@@ -21,14 +21,15 @@ import java.io.Serializable;
 public class Rune implements Serializable {
 
 	private static final long serialVersionUID = -5212321615648903989L;
+
 	private int count;
-	private long runeId;
+	private int runeId;
 
 	public int getCount() {
 		return count;
 	}
 
-	public long getRuneId() {
+	public int getRuneId() {
 		return runeId;
 	}
 }
